fix: fixed 'hyperswitch-keymanager' chart external postgresql enabled config
\#### Summary

This commit fixes a bug in the config template helper script that
requires the setting of `values.external.enabled: true` as opposed to
`values.external.postgresql.enabled: true` in order for the database
name to be populated.

\#### Testing

With this fix, the `create-keymanager-db` K8s job successfully
completes instead of failing due to the `pg_isready` command failing
because the DBNAME environment variable is empty.

---

Signed-Off-By: Willem Barkhuizen (<willem.barkhuizen@sanlam.co.za>)